Restore original srfi-43 vector API
This test was originally written to use the srfi-43 API for vectors.
This was changed in commit a1d30547c4a2d9021fdcbaef43b0e97fa65f013a
which changed all whitespace, which caused the API changes to be
invisible.

(rant "
I'm not sure why white-space changes are allowed. Don't people use
editors that can correctly handle whitespace? How is this even a
problem? Is this due to the popularity of some old 1980's editors
that haven't been ported over modern OS'es?
")