Restore the srfi-133 vector API
This undoes pull request #26

At the time of the above pull req, I did not reqlized that sfi-133
and srfi-43 were incompatible, while sharing the same function names.
I incorrectly assumed that this some strange gambit thing, rather than
a generic issue. So this pull req undoes that one, and provides a
commented-out shim for those of us who do not have srfi-133.

There's probaby a prettier way to implement the shim, but I don't
know what that is.